ui test: introduce SettingsAbsenceType page
