+++ src/main/resources/templates/settings/absence-types/settings_absence_types.html
@@ -38,7 +38,7 @@
                       th:text="#{action.save}"
                       type="submit"
                       class="button-main-green"
-                      data-test-id="settings-save-button"
+                      data-test-id="submit-custom-absence-types-button"
                       data-turbo="false"
                     ></button>
                     <button
